+++ src/cai_lab_to_nwb/zaki_2024/utils/__init__.py
@@ -1,6 +1,5 @@
 from .edf_slicing import get_session_slicing_time_range, get_session_run_time
-from .update_yaml_files import (
-    update_session_to_nwb_kwargs_yaml,
+from .source_data_path_resolver import (
     get_session_times_df,
     get_experiment_dir_path,
     get_edf_file_path,
@@ -11,3 +10,4 @@ from .update_yaml_files import (
     get_miniscope_folder_path,
     get_freezing_output_file_path,
 )
+from .update_session_to_nwb_kwargs_yaml import update_session_to_nwb_kwargs_yaml
